Add files via upload
This is about whether the person will pay the credit card bill or not. Based on the training of the dataset the result is predicted. The classification approach is applied for this problem as it is of 2-way classification business problem. For this business problem, the random forest algorithm is most suitable and it gives the accuracy of 80.25%.